Remove the defmacro-based API
This slightly prettier input format was actually pretty ugly when it
comes to the consistency.  Additionally the actually useful
function (sexp-to-xml) was weirdly named as a result.

Let's stick to accepting a regular list, either quoted or in a
variable (I mention it because the defmacro-based API didn't allow for
the latter).